+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fix submit button conditions not applying on the front end. ([#2568](https://github.com/verbb/formie/issues/2568))
 - Fix checkbox fields saving their default option values when explicitly unchecked on submit. ([#2457](https://github.com/verbb/formie/issues/2457))
 - Fix missing page-list validation indicators in the form builder when a form has no pages. ([#2786](https://github.com/verbb/formie/issues/2786))
+- Add a default country setting for Google Places address auto-complete fields. ([#2721](https://github.com/verbb/formie/issues/2721))
 
 ## 4.0.0-beta.2 - 2026-06-03